Replace deprecated subPlugin with `KotlinCompilerPluginSupportPlugin`
