--- kt/plugins/godot-gradle-plugin/src/main/resources/META-INF/services/org.jetbrains.kotlin.gradle.plugin.KotlinGradleSubplugin
@@ -1,1 +1,0 @@
-godot.gradle.subplugin.GodotSubPlugin
